Create blank.yml
README.md; Hello to all the people with great minds, 
minds of imagination, or just all those whos minds wonder all the time!! We are all intelligent in our own ways. Each and everyone of us are the same but different as far as how we think and what we know, and that's why I would like to get a website started called " Great Minds Intelligent times"!